backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
pdfminer.six==20221105
python-docx==1.1.0
nltk==3.8.1
python-dotenv==1.0.0
gunicorn==21.2.0

